Organize project in multiple folders
