+++ src/characters/adviser/types.ts
@@ -1,4 +1,4 @@
-import { Character } from "../character/types";
+import type { Character } from "../character/types";
 
 export interface Adviser extends Character {
   advisee: Character;
