+++ src/characters/characterTalk.ts
