+++ src/characters/king/types.ts
@@ -1,4 +1,4 @@
-import { Character } from "../character/types";
+import type { Character } from "../character/types";
 
 export interface King extends Character {
   yearsOfReign: number;
