+++ src/components/renderCard/renderCard.ts
@@ -1,5 +1,5 @@
-import type { CharacterTypes } from "../data.js";
-import { renderCardContainer } from "./renderCardContainer.js";
+import type { CharacterTypes } from "../../characters/data";
+import { renderCardContainer } from "../renderCardContainer/renderCardContainer.js";
 
 export const renderCard = (characters: CharacterTypes[]): HTMLElement => {
   const cardList = document.createElement("ul");
